requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
colorful==0.5.4
ety==1.3.3
Flask==1.1.2
future==0.18.2
futures==3.1.1
goslate==1.5.1
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Morfessor==2.0.6
numpy==1.19.2
PyDictionary==2.0.1
pymediawiki==0.6.4
pyunsplash==1.0.0rc2
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
treelib==1.6.1
textblob==0.15.3
urllib3==1.25.11
Werkzeug==1.0.1
